Made cleanup changes to large_drive.
